+++ library/src/main/java/com/getstream/sdk/chat/utils/MessageInputController.java
@@ -77,7 +77,7 @@ public class MessageInputController {
 
         binding.getRoot().setBackgroundResource(R.drawable.stream_round_thread_toolbar);
         binding.clTitle.setVisibility(View.VISIBLE);
-        binding.tvClose.setVisibility(View.VISIBLE);
+        binding.btnClose.setVisibility(View.VISIBLE);
 
         binding.clAddFile.setVisibility(View.GONE);
         binding.clCommand.setVisibility(View.GONE);
@@ -96,7 +96,7 @@ public class MessageInputController {
                 break;
             case COMMAND:
             case MENTION:
-                binding.tvClose.setVisibility(View.GONE);
+                binding.btnClose.setVisibility(View.GONE);
                 binding.clCommand.setVisibility(View.VISIBLE);
                 break;
         }
@@ -458,7 +458,7 @@ public class MessageInputController {
             closeCommandView();
         }else if (text.length() == 1) {
             onClickCommandViewOpen(text.startsWith("/"));
-        } else if (text.endsWith("@") && binding.clCommand.getVisibility() != View.VISIBLE) {
+        } else if (text.endsWith("@")) {
             onClickCommandViewOpen(false);
         } else {
             setCommandsMentionUsers(text);
@@ -490,20 +490,9 @@ public class MessageInputController {
                 binding.etMessage.setText("/" + ((Command) commands.get(position)).getName() + " ");
             else {
                 String messageStr = binding.etMessage.getText().toString();
-                String[] names = messageStr.split("@");
-                String messageStr_ = "";
-                int index;
-                if (messageStr.substring(messageStr.length() -1).equals("@"))
-                    index = names.length;
-                else
-                    index = names.length - 1;
-
-                for (int i = 0; i < index; i++){
-                    if (TextUtils.isEmpty(names[i])) continue;
-                    messageStr_ += "@" + names[i];
-                }
-                messageStr_ += "@";
-                binding.etMessage.setText(messageStr_ + ((User) commands.get(position)).getName() + " ");
+                String userName = ((User) commands.get(position)).getName();
+                String converted = StringUtility.convertMentionedText(messageStr, userName);
+                binding.etMessage.setText(converted);
             }
             binding.etMessage.setSelection(binding.etMessage.getText().length());
             closeCommandView();
@@ -531,7 +520,8 @@ public class MessageInputController {
             binding.tvCommand.setText(commandStr);
         } else {
             String[] names = string.split("@");
-            setMentionUsers(names[names.length - 1]);
+            if (names.length > 0)
+                setMentionUsers(names[names.length - 1]);
         }
     }