+++ library/src/main/java/com/getstream/sdk/chat/utils/StringUtility.java
@@ -75,8 +75,18 @@ public class StringUtility {
         if (bytes <= 0) return 0 + " B";
         if (bytes < unit) return bytes + " B";
         int exp = (int) (Math.log(bytes) / Math.log(unit));
-        String pre = String.valueOf(("KMGTPE").charAt(exp-1));
+        String pre = String.valueOf(("KMGTPE").charAt(exp - 1));
         DecimalFormat df = new DecimalFormat("###.##");
         return df.format(bytes / Math.pow(unit, exp)) + " " + pre + "B";
+    }
+
+    public static String convertMentionedText(String text, String userName) {
+        if (text.substring(text.length() -1).equals("@"))
+            return text + userName;
+
+        String[] names = text.split("@");
+        String last = names[names.length - 1];
+        return text.substring(0, text.length() - last.length()) + userName;
+
     }
 }