+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Upcoming
 
 - Fix Concurrent modification when removing member from channel
+- Fix automention input issue
 - Fix Sent message progress infinite
 - Fix channel delete event handling in ChannelList view model
 - Fix attachment duplicated issue when message edit